+++ src/Core/session.py
@@ -9,7 +9,7 @@ current_dir = os.path.dirname(__file__)
 parent_dir = os.path.abspath(os.path.join(current_dir, os.pardir))
 sys.path.append(parent_dir)
 
-from MainMenu import file_menu
+from Core import file_menu
 from Config import check
 
 def save_session(files):