+++ src/Core/template_menu.py
@@ -10,10 +10,10 @@ current_dir = os.path.dirname(__file__)
 parent_dir = os.path.abspath(os.path.join(current_dir, os.pardir))
 sys.path.append(parent_dir)
 
-from MainMenu import file_menu
-from MainMenu import edit_menu
+from Core import file_menu
+from Core import edit_menu
 from Config import check
-from MainMenu import themes
+from Core import themes
 
 template_window = None