+++ src/GUI/filetab.py
@@ -8,7 +8,7 @@ current_dir = os.path.dirname(__file__)
 parent_dir = os.path.abspath(os.path.join(current_dir, os.pardir))
 sys.path.append(parent_dir)
 
-from MainMenu import file_menu
+from Core import file_menu
 
 class ClosableTab(customtkinter.CTkFrame):
     def __init__(self, master, text, command, close_command, middle_click_command, *args, **kwargs):