+++ src/GUI/suggestions.py
@@ -9,8 +9,8 @@ parent_dir = os.path.abspath(os.path.join(current_dir, os.pardir))
 sys.path.append(parent_dir)
 
 from Config import check
-from MainMenu import file_menu
-from MainMenu import themes
+from Core import file_menu
+from Core import themes
 
 class Suggestions(tk.Frame):
     def __init__(self, master, root, *args, **kwargs):